+++ Packages/com.varneon.vudon.udonity/Runtime/Udon Programs/Internal/NotImplementedExceptionUtility.cs
@@ -15,10 +15,5 @@ namespace Varneon.VUdon.Udonity.Utilities
         {
             logger.LogError("Static Editor Flags are not implemented!");
         }
-
-        public void ThrowScaleToolNotImplementedException()
-        {
-            logger.LogError("Scale tool is not implemented yet, only visual preview is available");
-        }
     }
 }